--- seriesbr/bcb/json_to_df/__init__.py
@@ -1,1 +1,0 @@
-from . import series
