+++ seriesbr/ibge/__init__.py
@@ -1,1 +1,4 @@
-from .ibge import get_series, get_metadata
+from .series import get_series
+from .metadata import get_metadata
+
+__all__ = ['get_series', 'get_metadata']
