--- seriesbr/ibge/json_to_df/__init__.py
@@ -1,3 +1,0 @@
-from . import series
-
-__all__ = [series]
