+++ seriesbr/ibge/metadata.py
@@ -1,0 +1,28 @@
+from seriesbr.utils import requests
+
+
+def get_metadata(table):
+    """
+    Get an IBGE table metadata.
+
+    Examples
+    --------
+    >>> ibge.get_metadata(1419)
+                                                                 values
+    id                                                             1419
+    nome              IPCA - Variação mensal, acumulada no ano, acum...
+    URL                            http://sidra.ibge.gov.br/tabela/1419
+    pesquisa              Índice Nacional de Preços ao Consumidor Amplo
+    assunto                                           Índices de preços
+    periodicidade     {'frequencia': 'mensal', 'inicio': 201201, 'fi...
+    nivelTerritorial  {'Administrativo': ['N1', 'N6', 'N7'], 'Especi...
+    variaveis         [{'id': 63, 'nome': 'IPCA - Variação mensal', ...
+    classificacoes    [{'id': 315, 'nome': 'Geral, grupo, subgrupo, ...
+    """
+    url = build_url(table)
+    json = requests.get_json(url)
+    return json
+
+
+def build_url(table):
+    return f"https://servicodados.ibge.gov.br/api/v3/agregados/{table}/metadados"
