--- seriesbr/ibge/url_builders/__init__.py
@@ -1,1 +1,0 @@
-from . import metadata, series
