--- seriesbr/ipea/url_builders/metadata.py
@@ -1,2 +1,0 @@
-def build_url(code):
-    return f"http://ipeadata2-homologa.ipea.gov.br/api/v1/Metadados('{code}')", None
